Created The Ability to Destroy components
each entity now has the property alive which is checked before the entity is added to the newEntities List in the useGameLoop hook.
a new method has also been added to the entity class to turn the alive property false. this will remove it from the entity list.

Created the Component_Lifespan.
This allows entities to expire after a certain number of frames.